Improve integ tests
