+++ tests/integration/pilot/ingress_test.go
@@ -28,16 +28,13 @@ import (
 	"time"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	k8s "sigs.k8s.io/gateway-api/apis/v1beta1"
 
-	"istio.io/istio/pilot/pkg/model/kstatus"
 	"istio.io/istio/pkg/config/protocol"
 	"istio.io/istio/pkg/http/headers"
 	"istio.io/istio/pkg/test/echo/common/scheme"
 	"istio.io/istio/pkg/test/env"
 	"istio.io/istio/pkg/test/framework"
 	kubecluster "istio.io/istio/pkg/test/framework/components/cluster/kube"
-	"istio.io/istio/pkg/test/framework/components/crd"
 	"istio.io/istio/pkg/test/framework/components/echo"
 	"istio.io/istio/pkg/test/framework/components/echo/check"
 	"istio.io/istio/pkg/test/framework/components/environment/kube"
@@ -50,239 +47,6 @@ import (
 	helmtest "istio.io/istio/tests/integration/helm"
 	ingressutil "istio.io/istio/tests/integration/security/sds_ingress/util"
 )
-
-func TestGateway(t *testing.T) {
-	framework.
-		NewTest(t).
-		Run(func(t framework.TestContext) {
-			crd.DeployGatewayAPIOrSkip(t)
-
-			ingressutil.CreateIngressKubeSecret(t, "test-gateway-cert-same", ingressutil.TLS, ingressutil.IngressCredentialA,
-				false, t.Clusters().Configs()...)
-			ingressutil.CreateIngressKubeSecret(t, "test-gateway-cert-cross", ingressutil.TLS, ingressutil.IngressCredentialB,
-				false, t.Clusters().Configs()...)
-
-			retry.UntilSuccessOrFail(t, func() error {
-				err := t.ConfigIstio().YAML("", fmt.Sprintf(`
-apiVersion: gateway.networking.k8s.io/v1beta1
-kind: GatewayClass
-metadata:
-  name: istio
-spec:
-  controllerName: istio.io/gateway-controller
----
-apiVersion: gateway.networking.k8s.io/v1beta1
-kind: Gateway
-metadata:
-  name: gateway
-  namespace: istio-system
-spec:
-  addresses:
-  - value: istio-ingressgateway
-    type: Hostname
-  gatewayClassName: istio
-  listeners:
-  - name: http
-    hostname: "*.domain.example"
-    port: 80
-    protocol: HTTP
-    allowedRoutes:
-      namespaces:
-        from: All
-  - name: tcp
-    port: 31400
-    protocol: TCP
-    allowedRoutes:
-      namespaces:
-        from: All
-  - name: tls-cross
-    hostname: cross-namespace.domain.example
-    port: 443
-    protocol: HTTPS
-    allowedRoutes:
-      namespaces:
-        from: All
-    tls:
-      mode: Terminate
-      certificateRefs:
-      - kind: Secret
-        name: test-gateway-cert-cross
-        namespace: "%s"
-  - name: tls-same
-    hostname: same-namespace.domain.example
-    port: 443
-    protocol: HTTPS
-    allowedRoutes:
-      namespaces:
-        from: All
-    tls:
-      mode: Terminate
-      certificateRefs:
-      - kind: Secret
-        name: test-gateway-cert-same
----`, apps.Namespace.Name())).Apply()
-				return err
-			}, retry.Delay(time.Second*10), retry.Timeout(time.Second*90))
-			retry.UntilSuccessOrFail(t, func() error {
-				err := t.ConfigIstio().YAML(apps.Namespace.Name(), `
-apiVersion: gateway.networking.k8s.io/v1beta1
-kind: HTTPRoute
-metadata:
-  name: http
-spec:
-  hostnames: ["my.domain.example"]
-  parentRefs:
-  - name: gateway
-    namespace: istio-system
-  rules:
-  - matches:
-    - path:
-        type: PathPrefix
-        value: /get/
-    backendRefs:
-    - name: b
-      port: 80
----
-apiVersion: gateway.networking.k8s.io/v1alpha2
-kind: TCPRoute
-metadata:
-  name: tcp
-spec:
-  parentRefs:
-  - name: gateway
-    namespace: istio-system
-  rules:
-  - backendRefs:
-    - name: b
-      port: 80
----
-apiVersion: gateway.networking.k8s.io/v1beta1
-kind: HTTPRoute
-metadata:
-  name: b
-spec:
-  parentRefs:
-  - kind: Service
-    name: b
-  - name: gateway
-    namespace: istio-system
-  hostnames: ["b"]
-  rules:
-  - matches:
-    - path:
-        type: PathPrefix
-        value: /path
-    filters:
-    - type: RequestHeaderModifier
-      requestHeaderModifier:
-        add:
-        - name: my-added-header
-          value: added-value
-    backendRefs:
-    - name: b
-      port: 80
-`).Apply()
-				return err
-			}, retry.Delay(time.Second*10), retry.Timeout(time.Second*90))
-			for _, ingr := range istio.IngressesOrFail(t, t) {
-				t.NewSubTest(ingr.Cluster().StableName()).Run(func(t framework.TestContext) {
-					t.NewSubTest("http").Run(func(t framework.TestContext) {
-						paths := []string{"/get", "/get/", "/get/prefix"}
-						for _, path := range paths {
-							_ = ingr.CallOrFail(t, echo.CallOptions{
-								Port: echo.Port{
-									Protocol: protocol.HTTP,
-								},
-								HTTP: echo.HTTP{
-									Path:    path,
-									Headers: headers.New().WithHost("my.domain.example").Build(),
-								},
-								Check: check.OK(),
-							})
-						}
-					})
-					t.NewSubTest("tcp").Run(func(t framework.TestContext) {
-						_ = ingr.CallOrFail(t, echo.CallOptions{
-							Port: echo.Port{
-								Protocol:    protocol.HTTP,
-								ServicePort: 31400,
-							},
-							HTTP: echo.HTTP{
-								Path:    "/",
-								Headers: headers.New().WithHost("my.domain.example").Build(),
-							},
-							Check: check.OK(),
-						})
-					})
-					t.NewSubTest("mesh").Run(func(t framework.TestContext) {
-						_ = apps.A[0].CallOrFail(t, echo.CallOptions{
-							To:    apps.B,
-							Count: 1,
-							Port: echo.Port{
-								Name: "http",
-							},
-							HTTP: echo.HTTP{
-								Path: "/path",
-							},
-							Check: check.And(
-								check.OK(),
-								check.RequestHeader("My-Added-Header", "added-value")),
-						})
-					})
-					t.NewSubTest("status").Run(func(t framework.TestContext) {
-						retry.UntilSuccessOrFail(t, func() error {
-							gwc, err := t.Clusters().Kube().Default().GatewayAPI().GatewayV1beta1().GatewayClasses().Get(context.Background(), "istio", metav1.GetOptions{})
-							if err != nil {
-								return err
-							}
-							if s := kstatus.GetCondition(gwc.Status.Conditions, string(k8s.GatewayClassConditionStatusAccepted)).Status; s != metav1.ConditionTrue {
-								return fmt.Errorf("expected status %q, got %q", metav1.ConditionTrue, s)
-							}
-							return nil
-						})
-					})
-				})
-			}
-			t.NewSubTest("managed").Run(func(t framework.TestContext) {
-				t.ConfigIstio().YAML(apps.Namespace.Name(), `apiVersion: gateway.networking.k8s.io/v1beta1
-kind: Gateway
-metadata:
-  name: gateway
-spec:
-  gatewayClassName: istio
-  listeners:
-  - name: default
-    hostname: "*.example.com"
-    port: 80
-    protocol: HTTP
----
-apiVersion: gateway.networking.k8s.io/v1beta1
-kind: HTTPRoute
-metadata:
-  name: http
-spec:
-  parentRefs:
-  - name: gateway
-  rules:
-  - backendRefs:
-    - name: b
-      port: 80
-`).ApplyOrFail(t)
-				apps.B[0].CallOrFail(t, echo.CallOptions{
-					Port:   echo.Port{ServicePort: 80},
-					Scheme: scheme.HTTP,
-					HTTP: echo.HTTP{
-						Headers: headers.New().WithHost("bar.example.com").Build(),
-					},
-					Address: fmt.Sprintf("gateway-istio.%s.svc.cluster.local", apps.Namespace.Name()),
-					Check:   check.OK(),
-					Retry: echo.Retry{
-						Options: []retry.Option{retry.Timeout(time.Minute)},
-					},
-				})
-			})
-		})
-}
 
 func skipIfIngressClassUnsupported(t framework.TestContext) {
 	if !t.Clusters().Default().MinKubeVersion(18) {
